components/train/requirements.txt
tensorflow
scikit-learn
google-cloud-storage
